+++ packages/event-store-adapter-typeorm-v3/jest.config.js
@@ -1,7 +1,10 @@
 module.exports = {
+  testEnvironment: 'node',
   roots: ['<rootDir>/src'],
   transform: {
-    '^.+\\.tsx?$': 'ts-jest',
+    '^.+\\.tsx?$': ['ts-jest', {
+      tsconfig: '<rootDir>/tsconfig.test.json',
+    }],
   },
   moduleNameMapper: {
     '^@schemeless/(.*)$': '<rootDir>/../$1/src',