+++ packages/event-store/jest.config.js
@@ -1,14 +1,12 @@
 module.exports = {
   roots: ['<rootDir>/src'],
   transform: {
-    '^.+\\.tsx?$': 'ts-jest',
-  },
-  globals: {
-    'ts-jest': {
-      tsconfig: 'tsconfig.test.json',
-    },
+    '^.+\\.tsx?$': ['ts-jest', {
+      tsconfig: '<rootDir>/tsconfig.test.json',
+    }],
   },
   moduleNameMapper: {
     '^@schemeless/(.*)$': '<rootDir>/../$1/src',
   },
+  testEnvironment: 'node',
 };